fix query token
