+++ middleware/sso.go
@@ -15,12 +15,12 @@ func SSO(key, location string) gin.HandlerFunc {
 			c.Next()
 			return
 		}
-		queryToken := c.Query(key)
+		queryToken := c.Query("doorman_token")
 		if queryToken == "" {
 			cookieToken, err := c.Cookie(key)
 			if err != nil {
 				logrus.Warn("get cookie fail!", err)
-				url := c.Request.Proto + c.Request.Host + c.Request.RequestURI
+				url := c.Request.RequestURI
 				redirectUrl := base64.URLEncoding.EncodeToString([]byte(url))
 				c.Redirect(http.StatusFound, fmt.Sprintf("%s?redirectUrl=%s", location, redirectUrl))
 				c.Abort()
